pyproject.toml
[tool.poetry]
name = "fastapi_example"
version = "0.1.0"
description = ""
authors = ["IgorSemenov70 <87079215+IgorSemenov70@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
asyncpg = "^0.25.0"
uvicorn = "^0.17.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
pydantic = {extras = ["dotenv"], version = "^1.9.1"}
yoyo-migrations = "^7.3.2"
psycopg2 = "^2.9.3"
aiofiles = "^0.8.0"
aiohttp = "^3.8.1"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
